Fix compiler warnings
https://github.com/bibledit/cloud/issues/1095

+++ filter/mail.cpp
@@ -253,7 +253,7 @@ std::string filter_mail_limit_line_length_rfc5322(std::string body, const int le
     // in the range of the caret to the caret plus max line length.
     // Add a new line after that.
     // Update the caret positon and go to the next iteration.
-    if (auto gt_pos = body.rfind('>', caret + length);
+    if (auto gt_pos = body.rfind('>', caret + static_cast<size_t>(length));
         gt_pos != std::string::npos and gt_pos > caret)
     {
       body.insert(gt_pos + 1, "\n");
@@ -266,7 +266,7 @@ std::string filter_mail_limit_line_length_rfc5322(std::string body, const int le
     // in the range of the caret to the caret plus max line length.
     // Add a new line before that.
     // Update the caret position and go to the next iteration.
-    if (auto st_pos = body.rfind('<', caret + length);
+    if (auto st_pos = body.rfind('<', caret + static_cast<size_t>(length));
         st_pos != std::string::npos and st_pos > caret)
     {
       body.insert(st_pos, "\n");
@@ -278,7 +278,7 @@ std::string filter_mail_limit_line_length_rfc5322(std::string body, const int le
     // 1. Add a new line in the body at the range end.
     // 2. Update the caret position.
     if ((body.length() - caret) > static_cast<size_t>(length)) {
-      const size_t range_end = std::min(caret + length, body.length());
+      const size_t range_end = std::min(caret + static_cast<size_t>(length), body.length());
       body.insert(range_end, "\n");
       caret = range_end;
     }
